Send group members upon create group.
When a group is successfully created, send all members of the group to all members of the group. This means all members can keep track of the public keys and tokens of their peers. Therefore, they can send each other the symmetric key that is shared between them without the server knowing.