Migrated away from device groups.
Instead of using device groups, the server sends individual messages to each of the group members. This means that we no longer need to store a firebase id, we no longer need to make HTTP requests to the server, and we can encrypt group messages with the private key of the user receiving the message.
Tests have been fixed.
Refactore message type from JSON_UPDATE to GROUP_MESSAGE.